Add files via upload
This is the beginning of my code. I experimented a little bit with documentation I found online regarding support vector machines. My model training was not improving at all unless the data in the spreadsheet was scaled, so I used a code snipped from geeks4geeks to preprocess the data. After that the accuracy was changing when tuning hyperparameters. I experimented with trial and error, and reached an accuracy of up to 0.82